requirements.txt
streamlit
google-genai
pydot
Pillow
requests
python-dotenv